Another revamp to support pipelines soon.
